requirements.txt
Flask==3.1.0
tensorflow
keras
numpy>=1.26.0
opencv-python
Pillow==12.0.0
sqlite-utils==3.37
gunicorn==21.2.0